Pull out monad related things into TestContainers.Monad
