+++ testcontainers.cabal
@@ -32,6 +32,7 @@ library
     TestContainers.Docker
     TestContainers.Hspec
     TestContainers.Image
+    TestContainers.Monad
     TestContainers.Tasty
     TestContainers.Trace
